Deleting test_dot_product
